Move query classes to separate package
