+++ jdbc/src/main/java/tech/ydb/jdbc/query/JdbcLexer.java
@@ -1,4 +1,4 @@
-package tech.ydb.jdbc.common;
+package tech.ydb.jdbc.query;
 
 import java.util.function.Supplier;
